Split each DB models into each apps
