+++ apps/Client/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from .models import Client
+
+
+class ClientSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Client
+        fields = '__all__'
