+++ apps/Client/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import ClientListCreateAPIView
+
+
+urlpatterns = [
+    path('client/', ClientListCreateAPIView.as_view())
+]
