+++ apps/Division/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import DivisionListCreateAPIView
+
+
+urlpatterns = [
+    path('division/', DivisionListCreateAPIView.as_view())
+]
