+++ apps/Drawing/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from .views import DrawingListCreateAPIView
+
+
+urlpatterns = [
+    path('drawing/', DrawingListCreateAPIView.as_view())
+]
