--- apps/JIM/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import Client, Drawing, Division, Unit
-
-admin.site.register(Client)
-admin.site.register(Drawing)
-admin.site.register(Division)
-admin.site.register(Unit)
