--- apps/JIM/serializer.py
@@ -1,33 +1,0 @@
-from rest_framework import serializers
-from .models import Client, Unit,  Division, Drawing
-
-
-class ClientSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Client
-        fields = '__all__'
-
-
-class DivisionSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Division
-        fields = '__all__'
-
-
-class DrawingSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Drawing
-        fields = '__all__'
-
-
-class UnitSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Unit
-        fields = '__all__'
-
-    def get_material(self, obj):
-        return self.get_material_display()
