+++ apps/Part/serializers.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+from .models import Part
+
+
+class PartSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Part
+        fields = '__all__'
+
+    def get_material(self, obj):
+        return self.get_material_display()
