+++ apps/Part/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import PartListCreateAPIView
+
+urlpatterns = [
+    path('part/', PartListCreateAPIView.as_view())
+]
